Split `server.services.gateways` in two
